# Improve `word_wrap` performance