@flavorjones 2022-09-06
Is it worth considering inlining the `break_point` regex?

```ruby
def tweaked(text, line_width: 80, break_sequence: "\n")
  text.gsub(/(.{1,#{line_width}})(?:[^\n\S]+\n?|\n*\Z|\n)|\n/, "\\1#{break_sequence}").chomp!(break_sequence)
end
```

In the benchmark, this shows a (very) small improvement in speed and a further reduction in objects allocated, at the cost of readability.

```
Warming up --------------------------------------
              before     1.944k i/100ms
               after     4.276k i/100ms
             tweaked     4.406k i/100ms
Calculating -------------------------------------
              before     19.334k (± 3.2%) i/s -     97.200k in   5.032980s
               after     43.567k (± 2.5%) i/s -    218.076k in   5.008818s
             tweaked     44.612k (± 3.2%) i/s -    224.706k in   5.042143s

Comparison:
             tweaked:    44612.2 i/s
               after:    43567.1 i/s - same-ish: difference falls within error
              before:    19333.6 i/s - 2.31x  (± 0.00) slower

Calculating -------------------------------------
              before    13.993k memsize (     0.000  retained)
                       122.000  objects (     0.000  retained)
                        34.000  strings (     0.000  retained)
               after     5.595k memsize (     0.000  retained)
                        44.000  objects (     0.000  retained)
                        22.000  strings (     0.000  retained)
             tweaked     5.423k memsize (     0.000  retained)
                        41.000  objects (     0.000  retained)
                        20.000  strings (     0.000  retained)

Comparison:
             tweaked:       5423 allocated
               after:       5595 allocated - 1.03x more
              before:      13993 allocated - 2.58x more
```