@jonathanhefner 2022-09-06
@flavorjones  I debated that.  I didn't want the regex to be too cryptic, so I thought the small performance hit would be worth it.  But, on second thought, I think it is better to inline the `break_point` regex, and add ample comments via the `x` flag.  Please let me know if you think the readability could be improved.