@flavorjones 2022-09-06
I love this readability, but for some reason the commented version runs with a pretty big performance penalty? Not sure why and I haven't had a chance to dig in:

```ruby
def inline(text, line_width: 80, break_sequence: "\n")
  text.gsub(/(.{1,#{line_width}})(?:[^\n\S]+\n?|\n*\Z|\n)|\n/, "\\1#{break_sequence}").chomp!(break_sequence)
end

def inline_with_x(text, line_width: 80, break_sequence: "\n")
  text.gsub(
    %r/
      # Up to `line_width` characters
      ( .{1,#{line_width}} )
      # Followed by one of:
      #   (1) non-newline whitespace plus an optional newline
      #   (2) the end of the string, ignoring any trailing newlines
      #   (3) a newline
      (?: [^\n\S]+\n? | \n*\Z | \n )
      # OR an empty line
      | \n
    /x, "\\1#{break_sequence}"
  ).chomp!(break_sequence)
end
```

```
Calculating -------------------------------------
              before     16.212k (± 6.4%) i/s -     80.822k in   5.008038s
               after     34.270k (±11.9%) i/s -    172.592k in   5.119658s
              inline     36.748k (± 9.9%) i/s -    183.078k in   5.039354s
            inline/x     23.112k (± 9.5%) i/s -    115.872k in   5.065993s

Comparison:
              inline:    36747.6 i/s
               after:    34270.3 i/s - same-ish: difference falls within error
            inline/x:    23112.4 i/s - 1.59x  (± 0.00) slower
              before:    16212.5 i/s - 2.27x  (± 0.00) slower
```